Django_Chat/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-jsonform==2.23.2
django-model-utils==5.0.0
django-push-notifications==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
firebase-admin==6.7.0
frozenlist==1.6.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
http_ece==1.2.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
jsonfield==3.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
msgpack==1.1.0
multidict==6.4.3
packaging==24.2
pillow==11.2.1
pluggy==1.5.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
python-magic==0.4.27
pytz==2025.2
pywebpush==2.0.3
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9.1
service-identity==24.2.0
setuptools==79.0.0
six==1.17.0
sqlparse==0.5.3
swapper==1.4.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
whitenoise==6.9.0
yarl==1.20.0
zope.interface==7.2
